add a "prefix" column
The "org" always used to be the component of the url preceding the repo, which made
it useful in search_links templates. Now that the org can be specified via the config
file, that is no longer true. The new "prefix" column is now the component of the url
preceding the repo. If org is not specified in the config file, then prefix and org
will be the same value.