allow a unique name for the repo to be specified in the config file
It is possible for repos to have different hosts and/or paths, but the
same name. If this happens, the checkouts will attempt to use the same
directory, and one of them will fail.

This change allows a name to be added to the config file. This name
must be unique (per org), and will be used for the directory name of
the git checkout. If a name is not specified, the repo will be used
as the name.